feat(kube): new cloudsql config that is similar to the one from cloudRun
goal is to make the transition easier by sharing the db.

there is now a new config type for cloudsql that is similar to the one in cloudRun.

this also changes the default db names to be "fully-qualified" names, so those include full app name including environment

this helps to share the same db across multiple envs and even apps