feat: dotenv can now be enabled for local dev only
this is useful if you don't want that .env files are created during build, so that they don't accidentially show up in the build artifacts.

Having said that, the current build jobs only include usually dist in the artifacts,
so the .env files are not included in the build artifacts.
But this change is a precautionary measure in case we change that.

using dotEnv: true is still useful for (client) apps that rely on .env files during build-time, e.g. react-native