feat(pipeline): faster scheduler and job deployments
A query for existing jobs/schedulers is made and stored in a shell variable.

If the job/scheduler exists, the deployment will update right away, without trying to create.

This is cleaner because it doesn't require `set +e` to silence the potential pipeline error.
It will also run quicker, when the deployment deploys the same jobs and schedules again.

- [Job doing create fail update](https://git.panter.ch/impact-hub/impact-hub-api/-/jobs/3582299) - Duration: 3 minutes 30 seconds
- [Job doing direct updates, no fail](https://git.panter.ch/impact-hub/impact-hub-api/-/jobs/3582291) - Duration: 2 minutes 6 seconds

This review deployment has only one scheduler for testing.
The production deployment has 10 schedulers.
There will presumably be more time difference between the two on prod.