feat: custom env configs no longer merge arrays, but now optionally allow to specify an array
BREAKING CHANGE: custom env configs no longer merge arrays, but now optionally allow to specify an array

Background:

custom env config was merged with merging arrays as well.
this was mainly done because you sometimes add additional secrets to some envs (e.g. local or review).

however in some other cases, merging arrays was very unexpected and caused issues.
indeed we had even an example that showed this issue (custom-deploy).

the solution is to no longer merge arrays, but provide a new way of doing that.
the new way is to optionally allow that the value can be a function that takes the original value and returns the new value.