feat: remove envMode none
envMode none was confusing and led to unexpected behavior as it would merge jobs for all envs regardless of triggers.

the only reason it was there was to not run test and audit multiple times
when multiple envs are run at the same time. however this is only relevant for tagged release
where tests currently don't run at all.

we may revisit this in the future, but for now we remove it.